requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
APScheduler
asn1crypto
asttokens
attrs
beautifulsoup4
blinker
boto3
botocore
bs4
cachetools
certifi
cffi
charset-normalizer
click
contourpy
cryptography
cycler
decorator
dnspython
docstring_parser
dynaconf
email_validator
executing
fastapi
fastapi-cli
ffmpeg-python
fonttools
frozenlist
future
gitdb
GitPython
google-api-core
google-auth
google-cloud-aiplatform
google-cloud-alloydb-connector
google-cloud-bigquery
google-cloud-core
google-cloud-resource-manager
google-cloud-storage
google-crc32c
google-genai
google-resumable-media
googleapis-common-protos
greenlet
grpc-google-iam-v1
grpcio
grpcio-status
h11
httpcore
httptools
httpx
idna
ipython
ipython_pygments_lexers
jedi
Jinja2
jmespath
joblib
jsonschema
jsonschema-specifications
kiwisolver
m3u8
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mediapy
multidict
narwhals
numpy
packaging
pandas
parso
pexpect
pg8000
pgvector
pillow
prompt_toolkit
propcache
proto-plus
protobuf
psycopg2-binary
ptyprocess
pure_eval
pyarrow
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pydeck
Pygments
pyparsing
python-dateutil
python-dotenv
python-multipart
pytz
PyYAML
referencing
requests
rich
rpds-py
rsa
s3transfer
scikit-learn
scipy
scramp
shapely
shellingham
six
smmap
sniffio
soupsieve
SQLAlchemy
stack-data
starlette
streamlit
tenacity
threadpoolctl
toml
tornado
traitlets
typer
typing_extensions
tzdata
tzlocal
urllib3
uvicorn
uvloop
watchdog
watchfiles
wcwidth
websockets
yarl
